Shopping_Assistant/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
